+++ Sources/SwiftUIKit/Extensions/View/View+Enabled.swift
@@ -13,6 +13,9 @@ public extension View {
     /**
      Adds a condition that controls whether or not users can
      interact with this view.
+
+     This is just an inverted version of the `disabled` view
+     modifier, intended to increase readability.
      */
     func enabled(_ condition: Bool) -> some View {
         disabled(!condition)