+++ Sources/SwiftUIKit/Extensions/View/View+Visible.swift
@@ -11,7 +11,11 @@ import SwiftUI
 public extension View {
     
     /**
-     Hides the view if the provided `condition` is `true`.
+     Hides the view if the provided condition is `true`.
+
+     This function uses an `if` clause to toggle between the
+     hidden and non-hidden result types. Do only use it when
+     you know the implications of using conditional views.
      */
     @ViewBuilder
     func hidden(if condition: Bool) -> some View {
@@ -24,6 +28,13 @@ public extension View {
     
     /**
      Shows the view if the provided `condition` is `true`.
+
+     This is just an inverted `hidden` modifier, intended to
+     increase readability.
+
+     This function uses an `if` clause to toggle between the
+     hidden and non-hidden result types. Do only use it when
+     you know the implications of using conditional views.
      */
     func visible(if condition: Bool) -> some View {
         hidden(if: !condition)