+++ Sources/SwiftUIKit/Styles/ViewShadowStyle.swift
@@ -8,9 +8,6 @@
 
 import SwiftUI
 
-@available(*, deprecated, message: "ShadowStyle has been renamed to ViewShadowStyle to not conflict with the new SwiftUI 4 ShadowStyle.")
-public typealias ShadowStyle = ViewShadowStyle
-
 /**
  This struct represents a style with properties that matches
  the `SwiftUI`s `shadow` modifier.