+++ Release Notes.md
@@ -10,7 +10,11 @@ This will make the things in this library easier to understand and document, sin
 
 ## 4.0
 
-This major version removes all previously deprecated logic.
+This major version removes all previously deprecated logic to make the library tighter and cleaner.
+
+### 💥 Breaking Changes
+
+* The `Documents` namespace has been removed without being previously deprecated.